// src/app/modules/cms/components/menu/Model/MenuModal.ts
export interface MenuModal {
  parentId?: string
  name?: string
  nameNp?: string
  slug?: string
  iconType?: string
  menuIcon?: string
  menuImage?: string
  menuType?: string
  topMenu?: string
  featuredImage?: string
  featuredImageText?: string
  featuredImageTextNp?: string
  featuredImageUrl?: string
  status?: string | boolean
  menuLinkType?: string
  categoryId?: string
  contentId?: string
  url?: string
  file?: string
  serviceId?: string
  metaTitle?: string
  metaKeyword?: string
  metaDescription?: string
  headerImage?: string
  headerShortDescription?: string
  headerShortDescriptionNp?: string
  firstButtonText?: string
  firstButtonTextNp?: string
  firstButtonLink?: string
  secondButtonText?: string
  secondButtonTextNp?: string
  secondButtonLink?: string
  showNotificationBell?: boolean
  showToLoggedInUserOnly?: boolean
}